support for web app. UI Improvements
