+++ lib/components/forgotLoginPassword.dart
@@ -1,7 +1,8 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:pssswd/functions/app_logger.dart';
+
+import '../functions/app_logger.dart';
 
 class ForgotLoginPassword extends StatelessWidget {
   final _emailHandler = TextEditingController();
