--- lib/components/masterPasswordAck.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MasterPasswordAck extends StatelessWidget {
-  const MasterPasswordAck({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
