+++ lib/components/passwdCard.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:pssswd/functions/app_logger.dart';
 
+import '../functions/app_logger.dart';
 import '../screens/editEntryPage.dart';
 import '../functions/passwordDecrypter.dart';
 
@@ -45,13 +45,16 @@ class _PasswdCardState extends State<PasswdCard> {
 
     return Card(
       elevation: 2,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(5.0),
-            child: Container(
-              width: mediaQuery.size.width * 0.65,
+      child: Container(
+        width: (mediaQuery.size.width > 500) ? 800 : mediaQuery.size.width,
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(5.0),
+
+              // color: Colors.blue,
+              // width: mediaQuery.size.width * 0.65,
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
@@ -64,10 +67,14 @@ class _PasswdCardState extends State<PasswdCard> {
                       fontSize: 15,
                     ),
                   ),
-                  const Divider(
-                    thickness: 0.5,
-                    endIndent: 0,
-                    color: Color.fromARGB(88, 91, 91, 91),
+                  Container(
+                    width: (mediaQuery.size.width > 700)
+                        ? 600
+                        : mediaQuery.size.width * 0.6,
+                    child: Divider(
+                      thickness: 0.3,
+                      color: Color.fromARGB(88, 91, 91, 91),
+                    ),
                   ),
                   Text(
                     (widget.name.length > 15
@@ -77,66 +84,66 @@ class _PasswdCardState extends State<PasswdCard> {
                 ],
               ),
             ),
-          ),
-          Row(
-            children: [
-              IconButton(
-                onPressed: () async {
-                  try {
-                    final secureStorage = new FlutterSecureStorage();
-                    var masterPassword =
-                        await secureStorage.read(key: 'masterPassword');
+            Row(
+              children: [
+                IconButton(
+                  onPressed: () async {
+                    try {
+                      final secureStorage = new FlutterSecureStorage();
+                      var masterPassword =
+                          await secureStorage.read(key: 'masterPassword');
 
-                    var pss = PasswordDecrypter();
-                    final decryptedPassword = await pss.getDecryptedPassword(
-                        widget.password,
-                        widget.randForKeyToStore,
-                        widget.randForIV,
-                        masterPassword);
-                    Clipboard.setData(ClipboardData(text: decryptedPassword));
+                      var pss = PasswordDecrypter();
+                      final decryptedPassword = await pss.getDecryptedPassword(
+                          widget.password,
+                          widget.randForKeyToStore,
+                          widget.randForIV,
+                          masterPassword);
+                      Clipboard.setData(ClipboardData(text: decryptedPassword));
 
-                    Fluttertoast.showToast(
-                      msg: 'pssswd copied!',
-                      gravity: ToastGravity.CENTER,
-                    );
-                  } catch (e) {
-                    AppLogger.printErrorLog('Some error occured', error: e);
-                  }
-                },
-                icon: Icon(Icons.copy),
-              ),
-              IconButton(
-                onPressed: () async {
-                  try {
-                    var pss = PasswordDecrypter();
-                    var masterPassword =
-                        await secureStorage.read(key: 'masterPassword');
-                    final decryptedPassword = await pss.getDecryptedPassword(
-                        widget.password,
-                        widget.randForKeyToStore,
-                        widget.randForIV,
-                        masterPassword);
-                    var res = await Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => EditEntryPage(
-                          widget.name,
-                          widget.username,
-                          decryptedPassword,
-                          widget.url,
-                          widget.entry_id,
+                      Fluttertoast.showToast(
+                        msg: 'pssswd copied!',
+                        gravity: ToastGravity.CENTER,
+                      );
+                    } catch (e) {
+                      AppLogger.printErrorLog('Some error occured', error: e);
+                    }
+                  },
+                  icon: Icon(Icons.copy),
+                ),
+                IconButton(
+                  onPressed: () async {
+                    try {
+                      var pss = PasswordDecrypter();
+                      var masterPassword =
+                          await secureStorage.read(key: 'masterPassword');
+                      final decryptedPassword = await pss.getDecryptedPassword(
+                          widget.password,
+                          widget.randForKeyToStore,
+                          widget.randForIV,
+                          masterPassword);
+                      var res = await Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                          builder: (context) => EditEntryPage(
+                            widget.name,
+                            widget.username,
+                            decryptedPassword,
+                            widget.url,
+                            widget.entry_id,
+                          ),
                         ),
-                      ),
-                    );
-                  } catch (e) {
-                    AppLogger.printErrorLog('Some error occured', error: e);
-                  }
-                },
-                icon: Icon(Icons.edit),
-              ),
-            ],
-          )
-        ],
+                      );
+                    } catch (e) {
+                      AppLogger.printErrorLog('Some error occured', error: e);
+                    }
+                  },
+                  icon: Icon(Icons.edit),
+                ),
+              ],
+            ),
+          ],
+        ),
       ),
     );
   }
