+++ lib/providers/user_entries.dart
@@ -1,7 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:pssswd/functions/app_logger.dart';
+
+import '../functions/app_logger.dart';
 
 class UserEntries extends ChangeNotifier {
   var _entries;
