+++ pubspec.yaml
@@ -3,7 +3,7 @@ description: A cross-platform Password Management Application.
 
 publish_to: 'none' 
 
-version: 1.2.0+3
+version: 1.2.1+4
 
 environment:
   sdk: ">=2.16.2 <3.0.0"
